pr changes
fixed the sensitivity bug and made the recommended changes to the pr
