+++ Assets/GameAssets/Player/PlayerPrefabs/PlayerKit.prefab
@@ -470,7 +470,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bbe5baa3f4a28f24eb390f8394645708, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _GameplaySettingFilePath: /GameplaySettings.txt
+  _gameplaySettingFilePath: /GameplaySettings.txt
+  WasSettingsChanged: 1
 --- !u!1 &3991142726110527397
 GameObject:
   m_ObjectHideFlags: 0
