+++ Assets/General/Scripts/Settings/CameraSettings.cs.meta
