+++ Assets/StreamingAssets/GameplaySettings.txt
@@ -1,1 +1,1 @@
-105.1669 False False
+324.4561 False True
